aws/cloudformation/native/lambdas/image_embeddings_calculator/lambda_code/agent_image_embeddings_calculator/requirements.txt
Pillow==10.0.0
numpy===1.25.2
transformers==4.33.1
boto3==1.28.47
tritonclient[http]==2.36.0
sentry_sdk==1.31.0